move all resources to a single file
Also add some missing doc comments.
