--- raw.go
@@ -1,13 +1,0 @@
-package resource
-
-type RawID int
-
-type RawInfo struct {
-	Path string
-}
-
-type Raw struct {
-	ID RawID
-
-	Data []byte
-}
